chore: h2 기본 세팅을 위한 코드 변경

+++ src/main/java/com/shootdoori/match/dto/MatchApplicationResponseDto.java
@@ -4,7 +4,7 @@ import com.shootdoori.match.entity.MatchApplicationStatus;
 import java.time.LocalDateTime;
 
 public record MatchApplicationResponseDto(
-    Integer applicationId,
+    Long applicationId,
     Long applicantTeamId,
     Long targetTeamId,
     MatchApplicationStatus status
